add user id to apps table

+++ app/Http/Controllers/AppsController.php
@@ -6,23 +6,30 @@ use App\Http\Requests\CreateAppRequest;
 use App\Models\App;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 
 class AppsController extends Controller
 {
     public function index() : JsonResponse {
         return response()->json(
-            App::paginate(1000)
+            Auth::user()->apps()->paginate(1000)
         );
     }
 
     public function show(App $app) : JsonResponse {
+        
+        if (Auth::user()->cannot('view', $app)) {
+            abort(403);
+        }
+
         return response()->json($app);
     }
 
     public function store(CreateAppRequest $request, Str $stringSupport) : JsonResponse {
         return response()->json(
             App::create([
+                'user_id' => Auth::user()->id,
                 'app_name' => $request->get('app_name'),
                 'app_language_choice' => $request->get('app_language_choice'),
                 'app_id' => (string) $stringSupport->uuid(),
@@ -38,6 +45,10 @@ class AppsController extends Controller
     }
 
     public function destroy(App $app) : Response {
+        if (Auth::user()->cannot('delete', $app)) {
+            abort(403);
+        }
+
         $app->delete();
         return response()->noContent();
     }
